pyproject.toml
[tool.poetry]
name = "repl-nix-new-repl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
openpyxl = "^3.1.5"
python-pptx = "^1.0.2"
xhtml2pdf = "^0.2.16"
pypdf2 = "^3.0.1"
reportlab = "^4.2.2"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
